+++ doc/change-log.md
@@ -30,8 +30,9 @@
 - `TermCmd` 增加 `setParamsJson`, `setAckParamsJson` 方法
 - 新增 `QryTermCmdReq` 类
 - 新增 `info.gratour.jt808common.JT808Consts.ProtocolVersions.detectProtoVerBySimNoLen` 方法
-- `OpenStrmReq` 去除 `talkSendProtoVer` 属性
+- `OpenStrmReq` 和 `GnssOpenStrmReq` 去掉 `talkSendProtoVer` 属性 
 - `StrmStatus` 增加 `sent` 属性
+- 为了避免前端 js 解释时不丢失精度，`TermConnState` 的 `connAt`、`lastActive` 改为 String 类型
 
 ## 不兼容变更
 - `AvUploadReq` 增加 `fileSz` 必填属性